move files to respective version folders
