+++ lib/grape-swagger/openapi_3/doc_methods.rb
