+++ lib/grape-swagger/swagger_2/doc_methods.rb
