+++ lib/grape-swagger/swagger_2/endpoint.rb
